requirements.txt
uvicorn
fastapi
python-dateutil
pydantic
boto3
ocrmypdf
chromadb
langchain
langchain_community
langchain_huggingface
langchain_aws
langchain_experimental
langchain_core
accelerate
pypdf
unstructured
python-docx
python-multipart
pydantic[email]